@@ -7,7 +7,6 @@ import ch.alpine.owl.glc.core.GlcNode;
 import ch.alpine.owl.math.state.StateTime;
 import ch.alpine.tensor.Scalar;
 import ch.alpine.tensor.Tensor;
-import ch.alpine.tensor.ext.Lists;
 
 /** slightly different from {@link ImageCostFunction}
  * because evaluation only happens at last state of trajectory */
@@ -18,6 +17,6 @@ public final class SparseImageCostFunction extends ImageCostFunction {
 
   @Override // from CostIncrementFunction
   public Scalar costIncrement(GlcNode glcNode, List<StateTime> trajectory, Tensor flow) {
-    return flipYXTensorInterp.at(Lists.last(trajectory).state());
+    return flipYXTensorInterp.at(trajectory.getLast().state());
   }
 }
