@@ -19,18 +19,18 @@ import ch.alpine.owl.math.state.SimpleEpisodeIntegrator;
 import ch.alpine.owl.math.state.StateTime;
 import ch.alpine.owl.math.state.TrajectoryRegionQuery;
 import ch.alpine.owl.sim.LidarRaytracer;
-import ch.alpine.subare.core.RewardInterface;
-import ch.alpine.subare.core.StateActionCounter;
-import ch.alpine.subare.core.StepRecord;
-import ch.alpine.subare.core.td.Sarsa;
-import ch.alpine.subare.core.td.SarsaType;
-import ch.alpine.subare.core.util.DefaultLearningRate;
-import ch.alpine.subare.core.util.DiscreteQsa;
-import ch.alpine.subare.core.util.DiscreteStateActionCounter;
-import ch.alpine.subare.core.util.LearningRate;
-import ch.alpine.subare.core.util.PolicyBase;
-import ch.alpine.subare.core.util.PolicyType;
-import ch.alpine.subare.core.util.PolicyWrap;
+import ch.alpine.subare.api.LearningRate;
+import ch.alpine.subare.api.RewardInterface;
+import ch.alpine.subare.api.StateActionCounter;
+import ch.alpine.subare.api.StepRecord;
+import ch.alpine.subare.td.Sarsa;
+import ch.alpine.subare.td.SarsaType;
+import ch.alpine.subare.util.DefaultLearningRate;
+import ch.alpine.subare.util.DiscreteQsa;
+import ch.alpine.subare.util.DiscreteStateActionCounter;
+import ch.alpine.subare.util.PolicyBase;
+import ch.alpine.subare.util.PolicyType;
+import ch.alpine.subare.util.PolicyWrap;
 import ch.alpine.tensor.RealScalar;
 import ch.alpine.tensor.Scalar;
 import ch.alpine.tensor.Tensor;
