@@ -9,7 +9,7 @@ import ch.alpine.owl.util.ren.RegionRenders;
 import ch.alpine.owl.util.win.DemoInterface;
 import ch.alpine.owl.util.win.OwlAnimationFrame;
 import ch.alpine.sophus.math.api.Region;
-import ch.alpine.subare.core.td.SarsaType;
+import ch.alpine.subare.td.SarsaType;
 import ch.alpine.tensor.Tensor;
 import ch.alpine.tensor.Tensors;
 
