@@ -6,7 +6,6 @@ import java.util.List;
 import java.util.Map;
 
 import ch.alpine.tensor.Scalar;
-import ch.alpine.tensor.ext.Lists;
 
 public interface ObservingExpandInterface<T extends StateCostNode> extends ExpandInterface<T> {
   /** @return whether active or not
@@ -25,7 +24,7 @@ public interface ObservingExpandInterface<T extends StateCostNode> extends Expan
       observations.keySet().stream().findFirst().ifPresent(System.out::println);
       // ---
       List<Scalar> values = new ArrayList<>(observations.values());
-      System.out.println("cost gradient:\t" + values.get(0) + " -> " + Lists.last(values));
+      System.out.println("cost gradient:\t" + values.get(0) + " -> " + values.getLast());
     }
   }
 
