@@ -74,7 +74,7 @@ import ch.alpine.tensor.sca.Sign;
       for (int i = 0; i < samples.length(); i++) {
         ti = ti.add(spacing.Get(i));
         StateTime stateTime = new StateTime(samples.get(i), ti);
-        StateTime orig = Lists.last(trajectory).stateTime();
+        StateTime orig = trajectory.getLast().stateTime();
         // TODO OWL ALG this boolean expression appears twice => extract to function
         Tensor u = (transition instanceof DirectedTransition directedTransition && !directedTransition.isForward) //
             ? uBetween.apply(stateTime, orig) //
@@ -118,7 +118,7 @@ import ch.alpine.tensor.sca.Sign;
       Tensor points = Tensor.of(segment.stream().map(RrtsNode::state));
       Scalar maxLength = segment.stream().skip(1) //
           .map(node -> transitionSpace.connect(node.parent().state(), node.state()).length()).reduce(Max::of).get();
-      Scalar t0 = Lists.last(trajectory).stateTime().time();
+      Scalar t0 = trajectory.getLast().stateTime().time();
       int depth = Integers.log2Ceiling(Ceiling.FUNCTION.apply(maxLength.divide(Sign.requirePositive(dt))).number().intValue());
       if (!direction)
         points = Reverse.of(points);
@@ -132,7 +132,7 @@ import ch.alpine.tensor.sca.Sign;
       for (int i = 1; i < samples.length(); i++) {
         ti = ti.add(spacing.Get(i - 1));
         StateTime stateTime = new StateTime(samples.get(i), ti);
-        StateTime orig = Lists.last(trajectory).stateTime();
+        StateTime orig = trajectory.getLast().stateTime();
         Tensor u = direction //
             ? uBetween.apply(orig, stateTime) //
             : uBetween.apply(stateTime, orig);
