@@ -12,7 +12,6 @@ import ch.alpine.owl.math.state.TimeInvariantRegion;
 import ch.alpine.tensor.RealScalar;
 import ch.alpine.tensor.Scalar;
 import ch.alpine.tensor.Tensor;
-import ch.alpine.tensor.ext.Lists;
 import ch.alpine.tensor.nrm.Vector2Norm;
 import ch.alpine.tensor.sca.Sign;
 
@@ -33,7 +32,7 @@ public class RnNoHeuristicCircleGoalManager extends SimpleTrajectoryRegionQuery 
   @Override
   public Scalar costIncrement(GlcNode glcNode, List<StateTime> trajectory, Tensor flow) {
     StateTime from = glcNode.stateTime();
-    return Vector2Norm.between(from.state(), Lists.last(trajectory).state());
+    return Vector2Norm.between(from.state(), trajectory.getLast().state());
   }
 
   @Override
